// src/generated/client/src/models/Attachment.ts
export interface Attachment {
    /**
     * The date the object was published, in the site\'s timezone.
     * @type {Date}
     * @memberof Attachment
     */
    date?: Date;
    /**
     * The date the object was published, as GMT.
     * @type {Date}
     * @memberof Attachment
     */
    date_gmt?: Date;
    /**
     * 
     * @type {PostGuid}
     * @memberof Attachment
     */
    guid?: PostGuid;
    /**
     * Unique identifier for the object.
     * @type {number}
     * @memberof Attachment
     */
    id?: number;
    /**
     * URL to the object.
     * @type {string}
     * @memberof Attachment
     */
    link?: string;
    /**
     * The date the object was last modified, in the site\'s timezone.
     * @type {Date}
     * @memberof Attachment
     */
    modified?: Date;
    /**
     * The date the object was last modified, as GMT.
     * @type {Date}
     * @memberof Attachment
     */
    modified_gmt?: Date;
    /**
     * An alphanumeric identifier for the object unique to its type.
     * @type {string}
     * @memberof Attachment
     */
    slug?: string;
    /**
     * A named status for the object.
     * @type {string}
     * @memberof Attachment
     */
    status?: AttachmentStatusEnum;
    /**
     * Type of Post for the object.
     * @type {string}
     * @memberof Attachment
     */
    type?: string;
    /**
     * Permalink template for the object.
     * @type {string}
     * @memberof Attachment
     */
    permalink_template?: string;
    /**
     * Slug automatically generated from the object title.
     * @type {string}
     * @memberof Attachment
     */
    generated_slug?: string;
    /**
     * 
     * @type {PostTitle}
     * @memberof Attachment
     */
    title?: PostTitle;
    /**
     * The ID for the author of the object.
     * @type {number}
     * @memberof Attachment
     */
    author?: number;
    /**
     * Whether or not comments are open on the object.
     * @type {string}
     * @memberof Attachment
     */
    comment_status?: AttachmentCommentStatusEnum;
    /**
     * Whether or not the object can be pinged.
     * @type {string}
     * @memberof Attachment
     */
    ping_status?: AttachmentPingStatusEnum;
    /**
     * Meta fields.
     * @type {object}
     * @memberof Attachment
     */
    meta?: object;
    /**
     * The theme file to use to display the object.
     * @type {string}
     * @memberof Attachment
     */
    template?: string;
    /**
     * Alternative text to display when attachment is not displayed.
     * @type {string}
     * @memberof Attachment
     */
    alt_text?: string;
    /**
     * 
     * @type {AttachmentCaption}
     * @memberof Attachment
     */
    caption?: AttachmentCaption;
    /**
     * 
     * @type {AttachmentDescription}
     * @memberof Attachment
     */
    description?: AttachmentDescription;
    /**
     * Attachment type.
     * @type {string}
     * @memberof Attachment
     */
    media_type?: AttachmentMediaTypeEnum;
    /**
     * The attachment MIME type.
     * @type {string}
     * @memberof Attachment
     */
    mime_type?: string;
    /**
     * Details about the media file, specific to its type.
     * @type {object}
     * @memberof Attachment
     */
    media_details?: object;
    /**
     * The ID for the associated post of the attachment.
     * @type {number}
     * @memberof Attachment
     */
    post?: number;
    /**
     * URL to the original attachment file.
     * @type {string}
     * @memberof Attachment
     */
    source_url?: string;
}
// ...
export enum AttachmentStatusEnum {
    Publish = 'publish',
    Future = 'future',
    Draft = 'draft',
    Pending = 'pending',
    Private = 'private'
}
/**
* @export
* @enum {string}
*/
export enum AttachmentCommentStatusEnum {
    Open = 'open',
    Closed = 'closed'
}
/**
* @export
* @enum {string}
*/
export enum AttachmentPingStatusEnum {
    Open = 'open',
    Closed = 'closed'
}
/**
* @export
* @enum {string}
*/
export enum AttachmentMediaTypeEnum {
    Image = 'image',
    File = 'file'
}